+++ src/ULinkGame.Server/Actors/ActorDirectoryFeatures.cs
@@ -1,0 +1,8 @@
+using ULinkGame.Cluster;
+
+namespace ULinkGame.Server.Actors;
+
+public static class ActorDirectoryFeatures
+{
+    public static readonly ClusterFeature ActorDirectory = new("actor-directory");
+}